net/mlx5: fix VLAN PCP item calculation
The VLAN 16 bits tci field contains both values of PCP and VID. When
extracting any one of them - it is required not to affect the other one.
Previous to this commit in routine flow_dev_get_vlan_info_from_items()
we calcualted the PCP as follows:
    (1) vlan->vlan_tci &= MLX5DV_FLOW_VLAN_PCP_MASK;
    (2) vlan->vlan_tci |= <3 bits value of PCP>
In line (1) we should have used the negated mask ('~' operator) such
that only the PCP bits will be nullified before ORing them with the
updated PCP value.

Fixes: 9aee7a8418 ("net/mlx5: support push flow action on VLAN header")

Signed-off-by: Ophir Munk <ophirmu@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>