net/enetc: do not stall in clean Tx ring
Don't read the hardware CI register in a loop, read it once, clean up and
exit.
The issue with reading the register in a loop is that we're stalling here
trying to catch up with hardware which keeps sending traffic as long as it
has traffic to send, so in effect we could be waiting here for the Tx ring
to be drained by hardware, instead of us doing Rx in that meantime.
At the time we return the function there may be new BDs in the ring that
could be cleaned, we're just leaving those there for the next time.

Signed-off-by: Alex Marginean <alexandru.marginean@nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>