net/ice: fix hash flow segmentation fault
Macro rte_errno is not a static value, so it needs to be updated in all
error handling code.

Patch 'dc36bd5dfdeb' mistakenly consider that rte_errno is a constant,
which causes the unrecognized flow rule to be marked as recognition
success. Later, when the code tried to parse the flow rule, a null pointer
caused a segmentation fault.

Fixes: dc36bd5dfdeb ("net/ice: fix flow FDIR/switch memory leak")
Cc: stable@dpdk.org

Signed-off-by: Zhu Tao <taox.zhu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>