net/vmxnet3: fix RSS setting on v4
When calling to setup RSS on v4 API, ESX will expect
IPv4/6 TCP RSS to be set/requested mandatorily.

This patch will:
- Set IPv4/6 TCP RSS when these have not been set. A warning
message is thrown to make sure we warn the application we are
setting IPv4/6 TCP RSS when not set.
- An additional check has been added to dodge RSS configuration
altogether unless MQ_RSS has been requested, similar to v3.

The alternative (returning error) was considered, the intent
is to ease the task of setting up and running vmxnet3 in situations
where it's supposted to be most strightforward (testpmd, pktgen).

Open bug for this:
https://bugs.dpdk.org/show_bug.cgi?id=400

Signed-off-by: Eduard Serra <eserra@vmware.com>
Acked-by: Yong Wang <yongwang@vmware.com>
Fixes: 643fba770705 ("net/vmxnet3: v4 boot and guest UDP RSS configuration")
Signed-off-by: 0-day Robot <robot@bytheb.org>