net/bnxt: allow configuring vector mode
The bnxt driver has vector mode but it has limitations.
For example, rte_flow mark won't work in vector mode.
For this reason the user should be able to disable vector mode
as part of the config.

Make the configuration use the same as other drivers with
vector mode: ixge, i40e, ...

This will also make future support of vector mode on other
architectures possible.

Fixes: bc4a000f2f53 ("net/bnxt: implement SSE vector mode")
Cc: lance.richardson@broadcom.com
Cc: stable@dpdk.org
Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>