if_proxy: add example application
Add an example application showing possible library usage.
This is a simplified version of l3fwd where:
- many performance improvements has been removed in order to simplify
  logic and put focus on the proxy library usage,
- the configuration of forwarding has to be done by the user (using
  typical system tools on proxy ports) - these changes are passed to the
  application via library notifications.

It is meant to show how you can update some data from callbacks (routing
- see note below) and how those that are replicated (e.g. kept per
lcore) can be updated via event queueing (here neighbouring info).

Note: This example assumes that LPM tables can be updated by a single
writer while being used by others.  To the best of author's knowledge
this is the case (by preliminary code inspection) but DPDK does not make
such a promise.  Obviously, upon the change, there will be a transient
period (when some IPs will be directed still to the old destination) but
that is expected.

Note also that in some cases you might need to tweak your system
configuration to see effects.  For example you send Gratuitous ARP to
DPDK port and expect neighbour tables to be updated in application which
does not happen.  The packet will be sent to the kernel but it might
drop it, please check /proc/sys/net/ipv4/conf/dtap0/arp_accept and
related configuration options ('dtap0' here is just a name of your proxy
port).

Signed-off-by: Andrzej Ostruszka <aostruszka@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>