net/ice: add the DCF framework
The DCF (Device Config Function) feature works as a standalone PMD, its
hardware entity is the trusted VF with ID 0, and the software function
is to control the flow setting of other VFs by the mailbox with PF.

It doesn't handle packet Rx/Tx related things. This PMD needs the iAVF
virtchnl to send the ICE PF's AdminQ command.

Also for security, it needs to acquire the DCF capability from PF. The
flow related things will be added later, it shares most of the ICE flow
function. This mainly handles hardware initialization etc.

                               .-------------.
                             .-|             |
                           .-| |             |
                           | | |  iAVF PMD   |
                           | | |             |
                           | | |             |
                           | | '-------------'
                           | '-------------'
                           '-------------'
                          ^   Other VFs
                         /
                        / Flow Distribution
  .------------------. /
  |                  |/  DCF cap req   .-------------.
  |      CVL         .  <----------    |             |
  |     Kernel       |  ---------->    |             |
  .                  .   DCF cap rsp   |   DCF PMD   |
  |                  |                 |             |
  '------------------'  <-----------   |             |
                          rte_flow     '-------------'
                                             VF0

Signed-off-by: Haiyue Wang <haiyue.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>