lpm6: make IPv6 addresses immutable
None of the public functions modify IPv6 address passed.  So their
parameters are made const - with the exception of bulk functions.
This exception is due to compatibility problems - some compilers report
problems with const-casting of array of arrays.

Previously only lookup and add were updated to have addresses passed as
const so I'm adding this fixline.

Fixes: d82927d2f81d ("lpm6: make IPv6 address immutable")
Cc: stephen@networkplumber.org

Signed-off-by: Andrzej Ostruszka <aostruszka@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>