net/hns3: fix default error code of command interface
Currently, the hns3 PMD driver can interact with firmware through command
to complete hardware configuration. The driver calls internal interface
function named hns3_cmd_send to issues command to the firmware, and check
the execution result of the command through desc_ret returned by firmware
to driver.

As the design of error code, when device is resetting hns3_cmd_send will
only return -EBUSY or -EIO. But we found that if desc_ret is in [12,65535],
for example the item doesn't exist when issuing the command to query some
table item, hns3_cmd_send also return -EIO. This phenomenon will affect the
processing logic for the return value.

The root cause as below:
When desc_ret is in [12,65535], in the static functin named
hns3_cmd_convert_err_code called by hns3_cmd_send, matches the default case
and return -EIO. And then hns3_cmd_send return -EIO.

This patch fixes it with the following modification.
1. Change the return value of the default case in the static function named
   hns3_cmd_convert_err_code from -EIO to -EREMOTEIO.
2. Modify the comment add errcode description of the internal interface
   function named hns3_cmd_send.

Fixes: 737f30e1c3ab ("net/hns3: support command interface with firmware")
Cc: stable@dpdk.org

Signed-off-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: Wei Hu (Xavier) <xavier.huwei@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>