net/mlx5: fix zero value validation for metadata
MARK and META come and go to data path, so zero
value has special meaning to metadata actions/items.

The special meaning is to say that this packet don't
have META/MARK anymore - clear it.

Thus, metadata actions/items now allowed to set/match
on zero value as long as mask is not zero

Fixes: fcc8d2f716fd ("net/mlx5: extend flow metadata support")
Fixes: e554b672aa05 ("net/mlx5: support flow tag")
Fixes: 55deee1715f0 ("net/mlx5: extend flow mark support")
Cc: viacheslavo@mellanox.com
Cc: stable@dpdk.org

Signed-off-by: Wisam Jaddo <wisamm@mellanox.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>