net/vmxnet3: Added mtu_set() function to allow setting MTU.
(Picked up from @Charles Myers patch https://patchwork.dpdk.org/patch/57771/)

When the mtu_set() function is not implemented, rte_eth_dev_set_mtu()
fails with -ENOTSUP and mtu is not stored in the mtu field in the
rte_eth_dev_data.  This causes the mtu in Vmxnet3_MiscConf which is
shared with hypervisor to always be set to 1500.

This may cause issues receiving jumbo frames on Enhanced Data Path
N-VDS.

Signed-off-by: Eduard Serra <eserra@vmware.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>