eal/trace: implement trace memory allocation
Trace memory will be allocated per thread to enable lockless trace
events updates to the memory. The allocator will first attempt to
allocate from hugepage, if not available from hugepage then
fallback to malloc memory.

Later in the patches series, This API will be hooked to DPDK fastpath
and control plane thread creation API. It is possible for non
DPDK thread to use trace events, in that case, trace memory
will be allocated on the first event emission.

Signed-off-by: Jerin Jacob <jerinj@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>