ethdev: add tm cap for private shaper packet mode
Some NIC hardware have private shaper attached to
every node and has a limitation where packet mode is applied
both to the scheduling of a node's children using WFQ and
shaping of traffic out of the private shaper.
This cannot be expressed using existing capabilities or configurations.

So this patch adds a tm capability that if set by a PMD implies that
packet mode when configured is even applied to private shaper
connected to that node. This also implies the limitation
that all the SP children of that node should have same mode
at any point of time i.e either packet mode or byte mode and
same applies to private shaper in that NIC PMD.

This patch also adds missing capability that tells whether PMD
supports wfq weight mode or not.

Signed-off-by: Nithin Dabilpuram <ndabilpuram@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>