rte_ethdev: change rte_eth_dev_owner_get to return error if unowned
For applications that want to check if device is owned,
change the return value of rte_eth_dev_owner_get to return -ENOENT
if there is no owner for the device.

Change the two drivers (failsafe and netvsc) that are using
this experimental API to handle the new return value.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>