net/ena/base: fix testing for supported hash func
There was a bug in ena_com_fill_hash_function(), which was causing bit to
be shifted left one bit too much.

To fix that, the ENA_FFS macro is being ussed (returning the location of
the first bit set), hash_function value is being subtracted by 1 if any
hash function is supported by the device and BIT macro is used for
shifting for better verbosity.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>